Hotfix-133: Fix tzwin.py not importing when needed on Windows
