+++ dateutil/tz.py
@@ -14,10 +14,12 @@ import sys
 import os
 
 from six import string_types, PY3
+from dateutil.tzcommon import tzname_in_python2
 
 try:
     from dateutil.tzwin import tzwin, tzwinlocal
 except ImportError:
+    print 'import error'
     tzwin = tzwinlocal = None
 
 relativedelta = None
@@ -27,25 +29,8 @@ rrule = None
 __all__ = ["tzutc", "tzoffset", "tzlocal", "tzfile", "tzrange",
            "tzstr", "tzical", "tzwin", "tzwinlocal", "gettz"]
 
-
-def tzname_in_python2(namefunc):
-    """Change unicode output into bytestrings in Python 2
-
-    tzname() API changed in Python 3. It used to return bytes, but was changed
-    to unicode strings
-    """
-    def adjust_encoding(*args, **kwargs):
-        name = namefunc(*args, **kwargs)
-        if name is not None and not PY3:
-            name = name.encode()
-
-        return name
-
-    return adjust_encoding
-
 ZERO = datetime.timedelta(0)
 EPOCHORDINAL = datetime.datetime.utcfromtimestamp(0).toordinal()
-
 
 class tzutc(datetime.tzinfo):
